pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cl-monitoring"
version = "0.1.0"
description = "Local single-user read-only companion app for Crawlab"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "httpx>=0.27,<1",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.2,<3",
    "aiosqlite>=0.20,<1",
    "jinja2>=3.1,<4",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "coverage>=7,<8",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "types-PyYAML>=6.0,<7",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cl_monitoring", "src/integrations", "src/tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
